Changes: audio initializatino fix, touch button fix, Mirror window rendering options.
Fixes:

Audio:
If vr_forceOculusAudio was set to 1 and the rift audio device was disabled in windows, the game would fail to initialize the default audio device and would not start. The default audio device will now be initialized if the rift cannot be found.

Additionally, if vr_forceOculusAudio is set to 0, the audio device can now be specified via s_device.  Use 'listDevices' to see a list of available audio devices, and set 's_device' to set the number of the desired audio device. Setting s_device to -1 will use the default audio device.

Control fix:
The Touch thumb rest touch sensors can now be bound to controls in the game.  ( implemented but not working in previous build )

Changes:
Can now select the mirror window to display a single unwarped view, or the warped two eye view.  vr_stereoWindow 1 = stereo warped, vr_stereoWindow 0 = single unwarped.