Teleporting now works! But only with laser sight dot.
When aiming at a floor or surface you can stand on, the laser sight will change to a dot in a circle if you can teleport there, or a cross if you can't. Press the teleport button (I map it to B on Touch) to instantly teleport there if allowed. When aiming at anything else, it will be a dot.

It uses path-finding to only allow you to teleport to places you could reach by walking and jumping. There are a few bugs in the pathfinding though, which don't allow you to teleport past props like breakable stairs.