Use player clip instead of monster clip to build AAS.
I wasn't making a proper player AAS file before, because I was respecting the monster clip flag. Now the AAS file is much more accurate, and recognises almost anywhere the player can go. Almost all the previous problems are fixed.

Unfortunately, this makes the AAS file large.

There are a couple of new problems, where you think you should be able to walk, but can't. But they are minor.

Also it's easier for the player to cheat because some doors no longer stop AAS. I think we need to impose either a path segment limit or a path traveltime limit.