Triggered! Teleporting now triggers all events along path.
This includes taking damage, triggering NPCs to talk to you, opening doors, etc.

Known issues:
It doesn't always choose the best path. If there's a path that takes damage and a shortcut through a broken rail that doesn't (eg. Mars City Underground), it sometimes makes you take damage. But you can avoid that by teleporting to the edge of the broken rail first.
Sometimes the pathfinding gets stuck (when it needs to jump?) so part of the path is made straight instead.