Allow "The Lab" and "The Gallery" style teleport button.
Hold the teleport button to show teleport cursor, release to teleport.
There's also an option for the current method of using the gun sight when aiming at the ground, and just pressing to teleport.
The Left Hand and Head options are not implemented yet.

Also, teleporting now works without a gun.