Swapping hands in menu now swaps bindings. Add vr_handSwapsAnalogs.
It always makes sense to swap most buttons and triggers when swapping the weapon hand, because they are things like: fire weapon, reload weapon, change to fists, show PDA, show menu-PDA, and turn flashlight on or off, that directly affect the hand they are mapped to.

The analog sticks are less obvious, because they control movement and turning, in addition to changing weapons. So some people might like to swap the analog sticks when they swap weapon hands, and some won't.

Set vr_handSwapsAnalogs to false if you don't want to swap the analog sticks when you swap hands, or true if you do want to swap analog sticks. The default is true to match current behaviour, but I'm open to changing it.

This doesn't affect pressing in the analog sticks, because I use that for turning the flashlight on and off.