This is an interim commit, some changes ( teleport aiming ) will not function properly until pending pulls/merges are completed.
Changes:

Mod Directory:
All mod files moved to the 'Fully Possessed' directory.  This directory should exist in the same folder as the doom 3 executable. ( not in /base/).  Game will check for the file "Fully Possessed/guis/lookforward.tga" and fail with an error if unable to locate.

Teleport Aiming Beam:
1st implementation of parabolic aiming beam.  Aiming is not yet connected to the teleport mechanic until pending merges are completed.

Game window focus:
Mouse is now restrained to the game window while running to prevent focus loss while in game. Alt-tabbing to a different window will restore mouse movement.  This is experimental.