Updates to teleport aiming:
The code to check for headroom ( the please duck message ) would sometimes ask the player to duck if a model was in the way, when ducking would not make the location reachable. Added a clip check at crouch height to ensure the location is valid if crouching.

Removed 'please duck' text.  Updated the teleport model to turn red if crouching required.  Added a one shot tip box to inform  the player that a red teleport target means you need to crouch.