Hack to enable teleportation in final boss fight of standard campaign.
Switches the .aas file used when the player descends below threshold in level. This enables the player to teleport around the hellhole. There is still a small section of the level the player will need to walk - the path from the sarcophagus to the floor that descends to the lower level is not accessible in either aas file.
This hack is temporary until a better AAS file can be generated.  Also note that the uneven ground around the hellhole makes teleporting finicky.