Add VS2015 support and (theoretically) improve 2017 support. Fix readme (old ffmpeg download site is gone).
Note that there's a bug with CMake and Visual Studio 2015 / 2017 x64 on Windows 10 Creators Update where it tries to use an incomplete SDK by default that causes CMake to fail to get the version, so I had to manually specify an SDK to use on the command line.

Vista is needed for LCMapStringEx when compiling Voice.cpp on Visual Studio 2015 and above, so we set WINVER to vista on VS2015+.

And CodeAnalysis\SourceAnnotations.h fails to compile in regular C on VS2015+, so we only include it when building C++ files and not in tree.c

And we need to use the libs in the LibOVR VS2015 folder.

And we needed spaces around #defined macros in strings, but that was already done by others.