Automatically detect game path from Steam or GOG (Windows only).
The EXE (with DLLs) can now be placed anywhere and it will (probably) work. This is most useful when debugging from Visual Studio. The mod's "Fully Possessed" folder still needs to be in the game's folder though.

Partly based on GZDoom code by Randy Heit, Gaerzi, and Mike Swanson. But I made large improvements.
It can detect the Steam path even when not on the default Steam drive.