Dual wielding is working! Sort of.
The finger poses are wrong for one of the weapons, and the laser-sight is only on one weapon, and there's only one reload control, and the HUD and stat-watch only show whichever weapon is most powerful. But it's nearly working.
Both triggers work. Reaching behind your back allows you to cycle the weapon in whichever hand you use. The flashlight still works in either hand if it's empty.

player->weapon is now player->hands[0].weapon and player->hands[1].weapon;
player->flashlight is still player->flashlight and is in addition to the two weapons.

More code recognises that there are two weapons and affects both.
Some static variables had to be moved to the idPlayerHand class because the function is called for two different hands.
Some unused code is removed.
Some PDA handling is now tied to the hand holding the PDA.

Multiplayer GAME_RELIABLE_MESSAGE_CLIENT_HITSCAN_HIT is now (even more) broken.